+++ uui/src/lib/IconButton/IconButton.types.d.ts
@@ -1,7 +1,10 @@
+import type { BoxProps } from '$lib/Box/box.types';
+
 export type IconButtonColor = 'primary' | 'secondary' | 'inherit';
 
 export interface IconButtonProps
-  extends svelte.JSX.HTMLAttributes<HTMLButtonElement | HTMLAnchorElement> {
+  extends svelte.JSX.HTMLAttributes<HTMLButtonElement | HTMLAnchorElement>,
+    BoxProps<HTMLButtonElement | HTMLAnchorElement> {
   toggled?: boolean;
   iconClass?: string;
   toggleable?: boolean;
@@ -10,4 +13,5 @@ export interface IconButtonProps
   iconColor?: IconButtonColor;
   toggledIconContent?: string;
   toggledIconColor?: IconButtonColor;
+  ref?: HTMLButtonElement | HTMLAnchorElement;
 }