+++ uui/src/lib/index.ts
@@ -4,7 +4,6 @@ import Chip from './Chip/chip.svelte';
 import Step from './Stepper/Step.svelte';
 import List from './List/List.svelte';
 import Paper from './Paper/Paper.svelte';
-import Overlay from './Paper/Overlay.svelte';
 import Theme from './Theme/theme.svelte';
 import Button from './Button/Button.svelte';
 import Dialog from './Dialog/Dialog.svelte';
@@ -32,7 +31,6 @@ export {
   Step,
   List,
   Paper,
-  Overlay,
   Theme,
   Button,
   Dialog,
@@ -105,7 +103,7 @@ export type {
   TypographyPalette
 } from './shared/theme/typography-config/typography-config.types';
 export type { DialogActionsProps, DialogTitleProps } from './Dialog/Dialog.types';
-export type { IconButtonColor, IconButtonProps } from './IconButton/IconButton.types';
+export type { IconButtonProps } from './IconButton/IconButton.types';
 export type {
   StepVariant,
   StepOrientation,
@@ -132,4 +130,3 @@ export type {
   BreakpointValue
 } from './shared/theme/breakpoints/breakpoints.types';
 export type { PaperProps } from './Paper/Paper.types';
-export type { OverlayProps } from './Paper/Overlay.types';