+++ website/src/routes/paper/+page.svelte
@@ -1,7 +1,19 @@
 <script lang="ts">
   import VariantExample from '../../lib/VariantExample.svelte';
   import DocPage from '../../lib/doc-page.svelte';
-  import { Typography, Paper } from '@uui';
+  import { Typography } from '@uui';
+  import PaperTypesText from '../../../../uui/src/lib/Paper/Paper.types.d.ts?raw';
+  import PaperElevationExample from './paper-elevation-example.svelte';
+  import PaperElevationExampleText from './paper-elevation-example.svelte?raw';
+  import PaperSquareExample from './paper-square-example.svelte';
+  import PaperSquareExampleText from './paper-square-example.svelte?raw';
+  import PaperOutlinedExample from './paper-outlined-example.svelte';
+  import PaperOutlinedExampleText from './paper-outlined-example.svelte?raw';
+  import PaperColorExample from './paper-color-example.svelte';
+  import PaperColorExampleText from './paper-color-example.svelte?raw';
+  import InlineCode from '$lib/inline-code.svelte';
+  import PaperScrollExample from './paper-scroll-example.svelte';
+  import PaperScrollExampleText from './paper-scroll-example.svelte?raw';
 </script>
 
 <DocPage>
@@ -9,30 +21,15 @@
   <Typography variant="body1">Basic element for applying elevation to wrapped contents.</Typography>
   <Typography variant="h3">Props</Typography>
 
-  <VariantExample
-    code={`
-    export type PaperVariant = 'elevation' | 'outlined' | string;
-export interface PaperProps<T extends EventTarget> extends BoxProps<T> {
-  variant?: PaperVariant;
-  elevation?: number;
-  square?: boolean;
-}
-  `}
-  />
+  <VariantExample code={PaperTypesText.split('\n\n').pop() || ''} />
 
   <Typography variant="h4">Square</Typography>
   <Typography variant="body1">Default is rounded.</Typography>
 
   <div class="paper-example">
-    <div>
-      <Typography variant="subtitle1">False</Typography>
-      <Paper elevation={10}>10dp</Paper>
-    </div>
-    <div>
-      <Typography variant="subtitle1">True</Typography>
-      <Paper square elevation={10}>10dp</Paper>
-    </div>
+    <PaperSquareExample />
   </div>
+  <VariantExample code={PaperSquareExampleText} />
 
   <Typography variant="h4">Outlined</Typography>
   <Typography variant="body1">
@@ -41,80 +38,35 @@ export interface PaperProps<T extends EventTarget> extends BoxProps<T> {
   </Typography>
 
   <div class="paper-example">
-    <div>
-      <Typography variant="subtitle1">False</Typography>
-      <Paper elevation={0}>0dp</Paper>
-    </div>
-    <div>
-      <Typography variant="subtitle1">True</Typography>
-      <Paper variant="outlined" elevation={10}>0dp</Paper>
-    </div>
+    <PaperOutlinedExample />
   </div>
+  <VariantExample code={PaperOutlinedExampleText} />
 
   <Typography variant="h4">Elevation</Typography>
   <Typography variant="body1">Accepted elevation values range from 0 - 24.</Typography>
 
   <div class="paper-example">
-    <Paper elevation={0}>0dp</Paper>
-    <Paper elevation={1}>1dp</Paper>
-    <Paper elevation={2}>2dp</Paper>
-    <Paper elevation={3}>3dp</Paper>
-    <Paper elevation={4}>4dp</Paper>
-    <Paper elevation={5}>5dp</Paper>
-    <Paper elevation={6}>6dp</Paper>
-    <Paper elevation={7}>7dp</Paper>
-    <Paper elevation={8}>8dp</Paper>
-    <Paper elevation={9}>9dp</Paper>
-    <Paper elevation={10}>10dp</Paper>
-    <Paper elevation={11}>11dp</Paper>
-    <Paper elevation={12}>12dp</Paper>
-    <Paper elevation={13}>13dp</Paper>
-    <Paper elevation={14}>14dp</Paper>
-    <Paper elevation={15}>15dp</Paper>
-    <Paper elevation={16}>16dp</Paper>
-    <Paper elevation={17}>17dp</Paper>
-    <Paper elevation={18}>18dp</Paper>
-    <Paper elevation={19}>19dp</Paper>
-    <Paper elevation={20}>20dp</Paper>
-    <Paper elevation={21}>21dp</Paper>
-    <Paper elevation={22}>22dp</Paper>
-    <Paper elevation={23}>23dp</Paper>
-    <Paper elevation={24}>24dp</Paper>
+    <PaperElevationExample />
   </div>
 
-  <VariantExample
-    code={`
-<Paper  elevation={10}>10dp</Paper>
-<Paper  square elevation={10}>10dp</Paper>
-<Paper  elevation={10}>10dp</Paper>
-<Paper  variant="outlined" elevation={10}>10dp</Paper>
-<Paper  elevation={0}>0dp</Paper>
-<Paper  elevation={1}>1dp</Paper>
-<Paper  elevation={2}>2dp</Paper>
-<Paper  elevation={3}>3dp</Paper>
-<Paper  elevation={4}>4dp</Paper>
-<Paper  elevation={5}>5dp</Paper>
-<Paper  elevation={6}>6dp</Paper>
-<Paper  elevation={7}>7dp</Paper>
-<Paper  elevation={8}>8dp</Paper>
-<Paper  elevation={9}>9dp</Paper>
-<Paper  elevation={10}>10dp</Paper>
-<Paper  elevation={11}>11dp</Paper>
-<Paper  elevation={12}>12dp</Paper>
-<Paper  elevation={13}>13dp</Paper>
-<Paper  elevation={14}>14dp</Paper>
-<Paper  elevation={15}>15dp</Paper>
-<Paper  elevation={16}>16dp</Paper>
-<Paper  elevation={17}>17dp</Paper>
-<Paper  elevation={18}>18dp</Paper>
-<Paper  elevation={19}>19dp</Paper>
-<Paper  elevation={20}>20dp</Paper>
-<Paper  elevation={21}>21dp</Paper>
-<Paper  elevation={22}>22dp</Paper>
-<Paper  elevation={23}>23dp</Paper>
-<Paper  elevation={24}>24dp</Paper>
-  `}
-  />
+  <VariantExample code={PaperElevationExampleText} />
+
+  <Typography variant="h4">Color</Typography>
+  <Typography variant="body1">
+    The default backgroundColor of surface and color of on-surface can be changed using the
+    backgroundColor and color props with a string of
+    <InlineCode>type ThemeColors</InlineCode>
+  </Typography>
+  <div class="paper-example">
+    <PaperColorExample />
+  </div>
+  <VariantExample code={PaperColorExampleText} />
+
+  <Typography variant="h4">Additional Examples</Typography>
+
+  <Typography variant="h6">Scrolling</Typography>
+  <PaperScrollExample />
+  <VariantExample code={PaperScrollExampleText} />
 </DocPage>
 
 <style lang="scss">